feat: rework SchemaRow header
