+++ src/components/shared/index.ts
@@ -1,6 +1,5 @@
 export * from './Caret';
 export * from './Description';
 export * from './Format';
-export * from './Property';
 export * from './Types';
 export * from './Validations';
